requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
pydantic
httpx
pyttsx3
python-docx
pypdf